ai_eventstream/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Async Redis
redis[hiredis]==5.2.1
redis-om==0.3.3

# Redis Vector Search
redisvl==0.3.8

# Celery
celery[redis]==5.4.0
flower==2.0.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.7.0

# AI/ML
openai==1.59.7
tiktoken==0.8.0
sentence-transformers==3.3.1
numpy==2.2.1

# Optional: Google Gemini
google-generativeai==0.8.3

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
orjson==3.10.13

# Monitoring & Logging
prometheus-client==0.21.1
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.5
mypy==1.14.1

# CLI Tools
rich==13.9.4
